ic-ref: Separate out unit tests (dfinity-lab/ic-ref#49)
so far we ran the ic-ref internal unit tests (there is only one) as
part of `ic-ref-test`. That's odd and adds noise and an unnecessary
extra level of nesting to the report. So let’s just use a cabal test for
that.
Originally https://github.com/dfinity-lab/ic-ref/commit/50ae284a020093455dbbcabe29e7dfcc447237a7