ic-ref-(test): Implement start/stop/delete
and test it.

The testing is limited because there is no sensible reliable way to hold
back responses to inter-canister calls, so it is hard to test the
`stopping` state. This is part of future work.

Originally https://github.com/dfinity-lab/ic-ref/commit/b1a21378d81d443841912d3763576571d546b965