ic-ref(-test): Support and test ingress expiry
This implements (and tests) that requests in the past or far in the
future are rejected.

It does not implement actually forgetting about requests after their
expiry, nor does it test it – it is a bit pointless to test because
there is no _requirement_ that and when an implementation does that.

Originally https://github.com/dfinity-lab/ic-ref/commit/f3bc547473dfe1cc1d9327d053a5afb46cfc61a6