ic-ref: Introduce CanisterActions type
not used that, but this will be used both by the implementation of
certified variables, and of per-canister funds. So let’s introduce it
here so that these two PRs are easier to keep in in feature branches.

Originally https://github.com/dfinity-lab/ic-ref/commit/b9bd4d3f2e28ce966258eba1ba2392035edc54fa