ic-ref-test: Canister deletion tests
Avoid doing an update call to a non-existing canister; the test suite
will wait for a while to it to be ingressed. This allows us to run the
rest of the inter-canister tests against the replica.

Originally https://github.com/dfinity-lab/ic-ref/commit/824fd6d904b0635349820ca52262ce91fd9e695c