ic-ref: Process requests in the background
@hansl reported issues with `ic-ref` when doing a heavy request (such as
installing a large canister via a wallet canister): The subsequent
`/api/v1/read` would not be handled, and eventually timed out.

The fact that installing a large caniter via the wallet caniter is very
slow is hard to fix while we use an interpreter (optimizing the candid
library might help, though – the universal canister handles this use
case fine, as the newly added test shows).

But with some changes to the code (in `read`, do not try to read from an
`MVar` that is blocked, and make sure that the processing after
returning HTTP 202 happens in the background), the request status read
would be handled, and `dfx` keeps polling as expected.

Also added a `--http-log` option that prints some (not very helpful)
HTTP logs.

Originally https://github.com/dfinity-lab/ic-ref/commit/b0942527bfded26eacf6a1fca8ca77e708ba68fe