Squashed commit of the following:
this implements dfinity-lab/ic-ref#294.

As expected, this requires additional plumbing on the “agent” side, to
pass the effective canister id to all the places where it is needed.
This is probably indicative what is in store for the “real” agents out
there.

The tests that check the behavior of the IC if the effective id is
wrong or invalid are in its own group, so they can be skipped if the
replica initially is just fully permissive and doesn’t look at the URL.

Originally https://github.com/dfinity-lab/ic-ref/commit/215641629ace2d590013f89bb3f3fe3b0a0d3c3b