Memoize parseModule using uglymemo, to greatly remove memory footprint (#14)
Trying to run the `ic-ref-test` Test using Github Actions (https://github.com/dfinity/ic-hs/pull/2), I noticed that it failed actually running the test. The test runs for a while, and then `ic-ref` dies. My assumption was quickly that these free runners don't have enough memory. I confirmed this assumption by running `ic-ref +RTS -M5G` when running the test, and it fails rather early (after getting slower and slower while GHC keeps GCing more and more often.)

I did some heap profiling of `ic-ref`, and it seems that most of the memory is that of the parsed Wasm module. Now, `ic-ref-test` installs the same universal canister many many times. So I wrapped the `parseModule` function used a simple memo function, meaning that as long as the uploaded wasm module is the same, it will not parse it again, but re-use the existing parsed module. Now it needs less than 2GB, and judging from the profile, most of is bytestrings (so likely the frozen memory of the canisters).