Nix: Track haskell-upgrades (#37)
in #30 we tracked a PR against nixpkgs, which isn't very stable (e.g. if
it gets force-pushed to, it may break this repository). The changes we
are trying to include are still not in nixpkgs master, but they are
closer already, on the `haskell-upgrades` branch. So track that, until
it is merged into `master` (via https://github.com/NixOS/nixpkgs/pull/138596)