Configure mergify (#42)
this adds a mergify configuration to the repo, fixing #33.

TL;DR: Label a PR with `automerge-squash` and it will be merged once
approved an green. Also, mergify will update out-of-date branches as
`master` changes.

It is based on the `.mergify.yaml` from Motoko, but omitting the “fancy”
features (e.g. approving changelog changes and dependency bumps). We can
add them when and if we add auto dependency bumping, but we don’t have
that many dependencies, so maybe not needed.