[VER-1123] Implement metadata access (#55)
this is a reimplementation of @jwiegley’s #53; his PR gave me many ideas
that I feld would be very tedious and impolite to convey through prose
review, so I just wrote down how I would do it. I hope this achieves the
level of declaratively that we aim for (by moving plumbing code into
other modules, for example, and by having more precise types.)

In addition to #53, this performs more module validation, and adds
further tests, and fixes a bug where deleted canisters still show up in
the state tree (probably worth refactoring to avoid such bugs, but not
in this PR).