Remove duplicate compilation
We're noticing that on https://github.com/dfinity/ic-hs/pull/79 GHC
requires massive amounts of memory to build certain
modules (`IC.Test.Agent` being the worst offender requiring more than
13GB while GitHub Runners only have 7GB causing OOM kills).

What's worse is that multiple components in the cabal file require the
same modules causing the same module to be build multiple times, often
in parallel exacerbating the memory problem.

This fixes the duplicate compilation issue by letting the components
in the cabal file only depend on the ic-hs library.

This might be deterimental to ghcid as mentioned in the cabal
file. But the associated ticket has been closed so the problem is
hopefully fixed in an upcoming GHC release.