Restructure cabal file into a library with executables that depend on it
CI is way too slow and actually fails on darwin. The main reason for
the slowness is that each executable rebuilds all the modules from the
library.

This PR restructures the cabal file such that the executables depend
on the library instead of on the source code of the library. This
ensures the library is only built once.

The reason the cabal files was structured like this was to support
ghcid. However it seems GHCi supports loading multiple components
nowadays.