Restructure cabal file into a library with executables that depend on it (#141)
CI is way too slow and actually fails on darwin. The main reason for the slowness is that each executable rebuilds all the modules from the library.

This PR restructures the cabal file such that the executables depend on the library instead of on the source code of the library. This ensures the library is only built once.

The reason the cabal files was structured like this was to support ghcid. However most active maintainers are not using ghcid so we prefer to have faster builds instead.

Besides restructuring the cabal file this PR also switches building the haddocks from cabal to Nix which allows us to remove the `cabal.project` and `cabal.project.freeze` files which has the advantage that a `cabal build` will then pick the dependencies from the nix-shell instead of building them itself.