Strictify datatypes and state monads to reduce run-time memory
This sledgehammer approach is a quick experiment to see if we can reduce the 17GB memory usage of `ic-ref` when run inside the `ic-ref-test` and `coverage` jobs.

The right approach is to do a space profile and see what is being allocated and why. But I'm lazy so doing this first to see if it helps.

I'll close this draft PR if it doesn't help.